refactor: create separated files
