+++ src/ws_server/actions.ts
@@ -1,0 +1,7 @@
+export enum Action {
+  Register = "reg",
+  CreateRoom = "create_room",
+  AddUserToRoom = "add_user_to_room",
+  UpdateRoom = "update_room",
+  CreateGame = "create_game",
+};
